// src/app/teste/types.ts
export type JSONTYPES = {
    data: number
    historico_liquidacao: string
    valor: number
    data_do_recebimento: number
    valor_do_recebimento: number
    honorarios_repassados_via_outros_creditos: number
    custas: number
    recuperacao_de_prejuizo_ja_baixados_do_sistema: number
    desconto_concedido: number
    diferenca: number
  }
  